requirements.txt
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
pytest>=7.4.0
httpx>=0.25.0
PyJWT>=2.8.0
cryptography>=41.0.0
